pyproject.toml
[project]
name = "courswork-project-2"
version = "0.1.0"
description = ""
authors = [
    {name = "Ksenia",email = "kseniaver2014@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.2.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.17.0"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
# максимальная длина строки
line_length = 119
